Minor improvements to error handling. Having sporadic issues with cores not unloading or reloading properly. Trying to ignore status 400 at unload, to see if core is reloaded anyway.

+++ src/main/java/se/repos/testing/indexing/solr/TestIndexServerSolrHome.java
@@ -77,10 +77,17 @@ public abstract class TestIndexServerSolrHome {
 	}
 	
 	void destroy(File instanceDir) {
-		try {
-			FileUtils.deleteDirectory(instanceDir);
-		} catch (IOException e) {
-			throw new RuntimeException(e);
+		if (instanceDir == null) {
+			throw new IllegalArgumentException("instanceDir is null");
+		}
+		if (instanceDir.exists()) {
+			try {
+				FileUtils.deleteDirectory(instanceDir);
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+		} else {
+			logger.debug("Cleanup not needed because instanceDir " + instanceDir + " does not exist");
 		}
 	}
 	
@@ -99,6 +106,13 @@ public abstract class TestIndexServerSolrHome {
 		} catch (IOException e) {
 			throw new RuntimeException(e);
 		}
+		if (resources == null) {
+			throw new IllegalArgumentException("Failed to resolve resources with pattern " + pattern);
+		}
+		if (resources.length == 0) {
+			throw new IllegalArgumentException("No resources found in extraction pattern " + pattern);
+		}
+		logger.debug("Resource pattern {} matched {}", pattern, resources);
 		SortedMap<String, Resource> extract = new TreeMap<String, Resource>();
 		for (Resource r : resources) {
 			String full = r.getDescription();
